Ambient display music ticker [1/2]
This let us trigger an Ambient Display notification, even if Ambient
Display is disabled, so we can know the new track title/artist.
This is handy when the player starts playing a new track or even
when we use volume rocker skip track feature when screen is off.

We use provided Doze services to skip this when the proximity
sensor is covered or system Ambient events are pending.

If the feature is disabled, track info will be still shown by default
on keyguard and ambient when these modes are called by the system
or the user.

Change-Id: I1bf1ad83bce7c5a2a14824ac2993569d05356db8
Signed-off-by: Anirudh Gupta <anirudhgupta109@gmail.com>