Add a one time QS dialog for the caffeine tile
So why this dialog?

Because users are often confused by what this tile actually does. They can
figure out what the BT or WiFi tile is for but they can't figure out what this
one is for. To their credit, is not easily understood by what it does by the
name either. So instead of having to answer the question of 'what is this tile'
a million times, we're using what we have around us. We're adding a Dialog to
hopefully save us from having to answer that question.

Normally I would be against something like this, since we removed the data saver
dialog but I think this one is actually useful unlike the data saver one.

Change-Id: I3423a057baf660009b2a6cb7838fa5c8990e25c7